pyproject.toml
[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "Enterprise-grade document intelligence pipeline with agentic extraction"
authors = [{name = "FDE Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5.0",
    "pdfplumber>=0.11.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pymupdf>=1.23.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
layout = [
    "docling>=1.0.0",
    "layoutparser>=0.3.4",
]
vision = [
    "pdf2image>=1.16.0",
    "google-generativeai>=0.3.0",
    "Pillow>=10.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
all = [
    "docling>=1.0.0",
    "layoutparser>=0.3.4",
    "pdf2image>=1.16.0",
    "google-generativeai>=0.3.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']
skip-string-normalization = false

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short --cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*", "*/site-packages/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
